Implementation of pad for issue #4: allow specifying a pad (in seconds) with -p/--pad to buffer with some extra silence around subtitles with profanity. This is helpful especially in situations where the subtitle might be off by a fraction of a second or more and the silence would end too soon or start too late allowing it to slip through. The default is to not pad. I think using a value of 1 or 2 seconds would probably be best for this feature. Thanks for the idea @katanaBadger

+++ cleanvid.py
@@ -13,8 +13,15 @@ import delegator
 from subliminal import *
 from babelfish import Language
 from caselessdictionary import CaselessDictionary
+from itertools import tee
 
-__location__ = os.path.dirname(os.path.realpath(__file__))
+__script_location__ = os.path.dirname(os.path.realpath(__file__))
+
+# thanks https://docs.python.org/3/library/itertools.html#recipes
+def pairwise(iterable):
+  a, b = tee(iterable)
+  next(b, None)
+  return zip(a, b)
 
 ######## GetSubtitles #########################################################
 def GetSubtitles(vidFileSpec, srtLanguage):
@@ -67,27 +74,38 @@ class VidCleaner(object):
   tmpSubsFileSpec = ""
   outputVidFileSpec = ""
   swearsFileSpec = ""
+  swearsPadMillisec = 0
   swearsMap = CaselessDictionary({})
   muteTimeList = []
 
   ######## init #################################################################
-  def __init__(self, iVidFileSpec, iSubsFileSpec, oVidFileSpec, iSwearsFileSpec):
-    if os.path.isfile(iVidFileSpec):
+  def __init__(self, iVidFileSpec, iSubsFileSpec, oVidFileSpec, oSubsFileSpec, iSwearsFileSpec, swearsPadSec=0):
+
+    if (iVidFileSpec is not None) and os.path.isfile(iVidFileSpec):
       self.inputVidFileSpec = iVidFileSpec
     else:
       raise IOError(errno.ENOENT, os.strerror(errno.ENOENT), iVidFileSpec)
 
-    if os.path.isfile(iSubsFileSpec):
+    if (iSubsFileSpec is not None) and os.path.isfile(iSubsFileSpec):
       self.inputSubsFileSpec = iSubsFileSpec
 
-    if os.path.isfile(iSwearsFileSpec):
+    if (iSwearsFileSpec is not None) and os.path.isfile(iSwearsFileSpec):
       self.swearsFileSpec = iSwearsFileSpec
     else:
       raise IOError(errno.ENOENT, os.strerror(errno.ENOENT), iSwearsFileSpec)
 
-    self.outputVidFileSpec = oVidFileSpec
-    if os.path.isfile(self.outputVidFileSpec):
-      os.remove(self.outputVidFileSpec)
+    if (oVidFileSpec is not None) and (len(oVidFileSpec) > 0):
+      self.outputVidFileSpec = oVidFileSpec
+      if os.path.isfile(self.outputVidFileSpec):
+        os.remove(self.outputVidFileSpec)
+
+    if (oSubsFileSpec is not None) and (len(oSubsFileSpec) > 0):
+        self.cleanSubsFileSpec = oSubsFileSpec
+        if os.path.isfile(self.cleanSubsFileSpec):
+          os.remove(self.cleanSubsFileSpec)
+
+    self.swearsPadMillisec = swearsPadSec * 1000
+
 
   ######## del ##################################################################
   def __del__(self):
@@ -97,16 +115,14 @@ class VidCleaner(object):
       os.remove(self.tmpSubsFileSpec)
 
   ######## CreateCleanSubAndMuteList #################################################
-  def CreateCleanSubAndMuteList(self, cleanSubsFileSpec=None):
+  def CreateCleanSubAndMuteList(self):
     subFileParts = os.path.splitext(self.inputSubsFileSpec)
 
     self.tmpSubsFileSpec = subFileParts[0] + "_utf8" + subFileParts[1]
     shutil.copy2(self.inputSubsFileSpec, self.tmpSubsFileSpec)
     UTF8Convert(self.tmpSubsFileSpec)
 
-    if cleanSubsFileSpec is not None:
-      self.cleanSubsFileSpec = cleanSubsFileSpec
-    else:
+    if not self.cleanSubsFileSpec:
       self.cleanSubsFileSpec = subFileParts[0] + "_clean" + subFileParts[1]
 
     lines = []
@@ -125,20 +141,44 @@ class VidCleaner(object):
 
     subs = pysrt.open(self.tmpSubsFileSpec)
     newSubs = pysrt.SubRipFile()
-    for sub in subs:
+    newTimestampPairs = []
+
+    # for each subtitle in the set
+    # if text contains profanity...
+    # OR if the next text contains profanity and lies within the pad ...
+    # OR if the previous text contained profanity and lies within the pad ...
+    # then include the subtitle in the new set
+    prevNaughtySub = None
+    for sub, subPeek in pairwise(subs):
       newText = replacer.sub(lambda x: self.swearsMap[x.group()], sub.text)
-      if (newText != sub.text):
+      newTextPeek = replacer.sub(lambda x: self.swearsMap[x.group()], subPeek.text) if (subPeek is not None) else None
+           # this sub contains profanity, or
+      if ((newText != sub.text) or
+            # we have defined a pad, and
+          ((self.swearsPadMillisec > 0) and (newTextPeek is not None) and
+             # the next sub contains profanity and is within 5 seconds of this one, or
+           (((newTextPeek != subPeek.text) and ((subPeek.start.ordinal - sub.end.ordinal) <= self.swearsPadMillisec)) or
+             # the previous sub contained profanity and is within 5 seconds of this one
+            ((prevNaughtySub is not None) and ((sub.start.ordinal - prevNaughtySub.end.ordinal) <= self.swearsPadMillisec))))):
+        subScrubbed = (newText != sub.text)
         newSub = sub
         newSub.text = newText
         newSubs.append(newSub)
+        if subScrubbed:
+          prevNaughtySub = sub
+          newTimes = [pysrt.SubRipTime.from_ordinal(sub.start.ordinal - self.swearsPadMillisec).to_time(),
+                      pysrt.SubRipTime.from_ordinal(sub.end.ordinal + self.swearsPadMillisec).to_time()]
+        else:
+          prevNaughtySub = None
+          newTimes = [sub.start.to_time(), sub.end.to_time()]
+        newTimestampPairs.append(newTimes)
+      else:
+        prevNaughtySub = None
+
     newSubs.save(self.cleanSubsFileSpec)
 
-    newLines = []
-    for sub in newSubs:
-      newLines.append([sub.start.to_time(), sub.end.to_time()])
-
     self.muteTimeList = []
-    for timePair in newLines:
+    for timePair in newTimestampPairs:
       lineStart = (timePair[0].hour * 60.0 * 60.0) + (timePair[0].minute * 60.0) + timePair[0].second + (timePair[0].microsecond / 1000000.0)
       lineEnd = (timePair[1].hour * 60.0 * 60.0) + (timePair[1].minute * 60.0) + timePair[1].second + (timePair[1].microsecond / 1000000.0)
       self.muteTimeList.append("volume=enable='between(t," + format(lineStart, '.3f') + "," + format(lineEnd, '.3f') + ")':volume=0")
@@ -170,9 +210,10 @@ if __name__ == '__main__':
   parser.add_argument('-o', '--output', help='output video file', metavar='<output video>')
   parser.add_argument(      '--subs-output', help='output subtitle file', metavar='<output srt>', dest="subsOut")
   parser.add_argument('-w', '--swears', help='text file containing profanity (with optional mapping)', \
-                                        default=os.path.join(__location__, 'swears.txt'), \
+                                        default=os.path.join(__script_location__, 'swears.txt'), \
                                         metavar='<profanity file>')
   parser.add_argument('-l', '--lang',   help='language for srt download (default is "eng")', default='eng', metavar='<language>')
+  parser.add_argument('-p', '--pad',    help='pad (seconds) around profanity', metavar='<int>', dest="pad", type=int, default=0)
   args = parser.parse_args()
 
   inFile = args.input
@@ -186,8 +227,8 @@ if __name__ == '__main__':
     if (not subsFile):
       subsFile = GetSubtitles(inFile, lang)
 
-  cleaner = VidCleaner(inFile, subsFile, outFile, args.swears)
-  cleaner.CreateCleanSubAndMuteList(cleanSubsFileSpec=args.subsOut)
+  cleaner = VidCleaner(inFile, subsFile, outFile, args.subsOut, args.swears, args.pad)
+  cleaner.CreateCleanSubAndMuteList()
   cleaner.MultiplexCleanVideo()
 
 #################################################################################
